requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Click ~= 8.1.3
GitPython ~= 3.1.27
pandas ~= 1.5.2
py-markdown-table ~= 0.3.3
setuptools ~= 65.6.3
mistletoe ~= 1.0.1
